requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-environ==0.8.1
django-heroku==0.0.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
pykerberos==1.2.4
pymongo==3.12.1
python-snappy==0.6.1
pytz==2021.3
sqlparse==0.2.4
whitenoise==6.0.0
psycopg2